♻️ Adapt import statements
